CitationCheck2/old/requirements.txt
pywebview==5.2